Minor component code refactor
