+++ src/components/AudioPlayer/AudioPlayer.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const AudioPlayer = (props) => {
+    return(
+        <div className="player">
+            {props.children}
+        </div>
+    );
+}
+
+export default AudioPlayer;
