+++ src/components/Cards/index.jsx
@@ -1,0 +1,1 @@
+export {default as ProductCard} from './ProductCard';
